+++ supplier_server_mock/supplier_server/constants.py
@@ -6,6 +6,9 @@ PRODUCTS = [
         'is_refundable': True,
         'cutoff_time': 24,
         'ticket_content_type': 'CODE128',
+        'max_tickets_per_order': 10,
+        'required_visitor_data': [],
+        'required_order_data': [],
         # but this is not part of spec...
         'cancelled_bookings': []
     },
@@ -17,6 +20,9 @@ PRODUCTS = [
         'is_refundable': False,
         'cutoff_time': 0,
         'ticket_content_type': 'CODE128',
+        'max_tickets_per_order': 10,
+        'required_visitor_data': ['FULL_NAME', 'PHONE'],
+        'required_order_data': ['PICKUP_LOCATION', 'PASSPORT_ID'],
         'cancelled_bookings': []
     },
     {
@@ -27,6 +33,9 @@ PRODUCTS = [
         'is_refundable': True,
         'cutoff_time': 0,
         'ticket_content_type': 'CODE128',
+        'max_tickets_per_order': 25,
+        'required_visitor_data': [],
+        'required_order_data': ['PICKUP_LOCATION', 'PASSPORT_ID', 'FLIGHT_NUMBER'],
         'cancelled_bookings': []
     },
     {
@@ -37,6 +46,7 @@ PRODUCTS = [
         'is_refundable': True,
         'cutoff_time': 10,
         'ticket_content_type': 'PDF',
+        'required_visitor_data': ['EMAIL', 'DATE_OF_BIRTH'],
         'cancelled_bookings': []
     },
     {
@@ -46,8 +56,13 @@ PRODUCTS = [
         'is_refundable': False,
         'cutoff_time': 0,
         'ticket_content_type': 'CODE128',
+        'max_tickets_per_order': 5,
+        'required_order_data': ['NATIONALITY'],
         'cancelled_bookings': []
     },
 ]
 VARIANTS = ('Adult', 'Child')
 MAX_DATE_RANGE = 6  # in months
+PRODUCTS_CURRENCIES = {
+    'A400-FX': 'USD',
+}